Fix five real bugs found by deep audit: buffer reset, nop offset, bytes_count overflow, address OOB, malloc NULL check
Agent-Logs-Url: https://github.com/dgatf/I2C-slave-multi-address-RP2040/sessions/69af45ec-082e-4bf2-8758-7e3f3b8d9925

+++ sdk/main.c
@@ -51,7 +51,7 @@ void i2c_request_handler(uint8_t address) {
     }
 }
 
-void i2c_stop_handler(uint8_t length) { printf("\nTotal bytes: %u", length); }
+void i2c_stop_handler(uint16_t length) { printf("\nTotal bytes: %u", length); }
 
 int main() {
     stdio_init_all();
